move test_time to apptest_time, remove deprecated time.clock
